(MAINT) Allow parallel acceptance tests runs
This adds a new rake task under parallel:acceptance which runs each spec
file in a separate process, more than halving the runtime for the entire
suite.

Note that for the moment this is only available locally. Running under
vagrant doesn't work - as the beaker provisioner trips up by trying to
launch multiple vagrant boxes with the same name. The pooler provisioner
does work (it launches multiple sets of machines) however this doesn't
bring down the runtime.

* The pooler spends a lot more time setting up the VM and installing PE
  that it does running the tests.
* It triggers the download and upload of the very large PE tarball once
  per spec file, which increases the runtime
* It might be worth coming back to when the above point is resolved, or
  when beaker is able to use snapshots of already running PE installs
