+++ .rspec_parallel
@@ -1,0 +1,2 @@
+--format ParallelTests::RSpec::RuntimeLogger
+--format ParallelTests::RSpec::SummaryLogger
