+++ spec/support/context/destroy_created_resources.rb
@@ -1,0 +1,5 @@
+shared_context 'destroys created resources after use' do
+  after(:all) do
+    @client.destroy_virtual_machine(@machine)
+  end
+end
